Move open_live()
